+++ spark-connector/spark-common/src/main/java/org/apache/gravitino/spark/connector/hive/SparkHiveTable.java
@@ -25,6 +25,7 @@ import org.apache.gravitino.spark.connector.PropertiesConverter;
 import org.apache.gravitino.spark.connector.SparkTransformConverter;
 import org.apache.gravitino.spark.connector.SparkTypeConverter;
 import org.apache.gravitino.spark.connector.utils.GravitinoTableInfoHelper;
+import org.apache.gravitino.spark.connector.utils.HiveGravitinoOperationOperator;
 import org.apache.kyuubi.spark.connector.hive.HiveTable;
 import org.apache.kyuubi.spark.connector.hive.HiveTableCatalog;
 import org.apache.spark.sql.SparkSession;
@@ -40,6 +41,7 @@ import org.apache.spark.sql.types.StructType;
 public class SparkHiveTable extends HiveTable implements SupportsPartitionManagement {
 
   private GravitinoTableInfoHelper gravitinoTableInfoHelper;
+  private HiveGravitinoOperationOperator hiveGravitinoOperationOperator;
 
   public SparkHiveTable(
       Identifier identifier,
@@ -58,6 +60,7 @@ public class SparkHiveTable extends HiveTable implements SupportsPartitionManage
             propertiesConverter,
             sparkTransformConverter,
             sparkTypeConverter);
+    this.hiveGravitinoOperationOperator = new HiveGravitinoOperationOperator(gravitinoTable);
   }
 
   @Override
@@ -84,29 +87,29 @@ public class SparkHiveTable extends HiveTable implements SupportsPartitionManage
   @Override
   public void createPartition(InternalRow ident, Map<String, String> properties)
       throws PartitionAlreadyExistsException, UnsupportedOperationException {
-    gravitinoTableInfoHelper.createPartition(ident, properties, partitionSchema());
+    hiveGravitinoOperationOperator.createPartition(ident, properties, partitionSchema());
   }
 
   @Override
   public boolean dropPartition(InternalRow ident) {
-    return gravitinoTableInfoHelper.dropPartition(ident, partitionSchema());
+    return hiveGravitinoOperationOperator.dropPartition(ident, partitionSchema());
   }
 
   @Override
   public void replacePartitionMetadata(InternalRow ident, Map<String, String> properties)
       throws NoSuchPartitionException, UnsupportedOperationException {
-    throw new UnsupportedOperationException("Partition replace is not supported");
+    throw new UnsupportedOperationException("Replace partition is not supported");
   }
 
   @Override
   public Map<String, String> loadPartitionMetadata(InternalRow ident)
       throws UnsupportedOperationException {
-    return gravitinoTableInfoHelper.loadPartitionMetadata(ident, partitionSchema());
+    return hiveGravitinoOperationOperator.loadPartitionMetadata(ident, partitionSchema());
   }
 
   @Override
   public InternalRow[] listPartitionIdentifiers(String[] names, InternalRow ident) {
-    return gravitinoTableInfoHelper.listPartitionIdentifiers(names, ident, partitionSchema());
+    return hiveGravitinoOperationOperator.listPartitionIdentifiers(names, ident, partitionSchema());
   }
 
   @Override
@@ -114,12 +117,17 @@ public class SparkHiveTable extends HiveTable implements SupportsPartitionManage
     if (n == 0) {
       return gravitinoTableInfoHelper;
     }
+
+    if (n == 1) {
+      return hiveGravitinoOperationOperator;
+    }
+
     throw new IndexOutOfBoundsException("Invalid index: " + n);
   }
 
   @Override
   public int productArity() {
-    return 1;
+    return 2;
   }
 
   @Override
